+++ CHANGELOG.md
@@ -19,3 +19,4 @@ Changed for individual apps are listed in `apps/appname/ChangeLog`
 * If `bin/pre-publish.sh` has been run and recent.csv created, add 'Sort By' chip
 * New 'espruinotools' which fixes pretokenise issue when ID follows ID (fix #416)
 * Improve upload of binary files
+* App description can now be markdown